backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
sqlalchemy==2.0.22
asyncpg==0.30.0
psycopg2-binary==2.9.10
greenlet==3.1.1
langchain==0.3.25
langchain-openai==0.2.14
langchain-core==0.3.58
langgraph==0.2.60
python-multipart==0.0.12
PyPDF2==3.0.1
python-dotenv==1.0.0
pgvector==0.2.5

# High-volume infrastructure dependencies
redis==5.0.1
elasticsearch==8.11.0
celery==5.3.4
